Fix initial cleanup

+++ testing/libs/dataset_testing_lib.py
@@ -69,13 +69,15 @@ class DatasetTestingLibrary:
         """Recursively remove files from directory_to_erase."""
         files_to_erase = glob.glob(f"{str(directory_to_erase)}/**", recursive=True)
         dirs_to_erase = []
+        print(files_to_erase)
 
         # Remove files first
         for file in files_to_erase:
-            if not os.path.isdir(file):
-                os.remove(file)
-            else:
-                dirs_to_erase.append(os.path.join(file))
+            if os.path.exists(file):
+                if not os.path.isdir(file):
+                    os.remove(file)
+                else:
+                    dirs_to_erase.append(os.path.join(file))
         # Sort longest to shortest to delete child directories first
         dirs_to_erase = sorted(dirs_to_erase, key=len, reverse=True)
